Change id type to long

+++ src/main/java/dmax/words/domain/Persistable.java
@@ -6,8 +6,8 @@ package dmax.words.domain;
  */
 public interface Persistable {
 
-    int getId();
+    long getId();
 
-    void setId(int id);
+    void setId(long id);
 
 }
